- Bug #579642 fixed. - A couple of other small bugs connected to the change to package fix. Beta testers: please, check that everything still works as good as before rev 175! - Setup.py doesn't try to remove /usr/bin/dockbarx.py unless it's called with the command install. - Small bug in setup.py fixed.

+++ setup.py
@@ -2,9 +2,7 @@
 
 from distutils.core import setup
 import os
-
-print "Installing DockbarX."
-print
+import sys
 
 setup(name='Dockbarx',
       version='0.3.1',
@@ -23,11 +21,11 @@ setup(name='Dockbarx',
 
 
 # Remove old dockbarx.py so that it isn't imported instead of the package dockbarx when dockbarx is run.
-if os.path.exists('/usr/bin/dockbarx.py'):
+if os.path.exists('/usr/bin/dockbarx.py') and len(sys.argv) == 2 and sys.argv[1] == 'install':
     print
     print 'There is a dockbarx.py in /usr/bin. This has to be removed to make DockbarX run correctly.'
-    remove = raw_input('Remove /usr/bin/dockbarx.py? (y/n)')
-    if remove[0].lower() == 'y':
+    remove = raw_input('Remove /usr/bin/dockbarx.py? (Y/n)')
+    if remove == ""or remove[0].lower() == 'y':
         os.remove('/usr/bin/dockbarx.py')
     else:
         print '/usr/bin/dockbarx.py is not removed. Please remove it or rename it manually.'
